+++ objects/components/Transform.h
@@ -8,24 +8,24 @@
 class Transform : public Component
 {
 public:
-  Transform(float xPos, float yPos, float width, float height);
+  Transform(double xPos, double yPos, double width, double height);
 
-  void fixedUpdate(float dt) override;
+  void fixedUpdate(double dt) override;
 
-  float getX() const;
-  float getY() const;
+  double getX() const;
+  double getY() const;
 
-  void move(Vec2<float> vector);
+  void move(Vec2<double> vector);
 
-  std::vector<Vec2<float>> getMesh() const;
+  std::vector<Vec2<double>> getMesh() const;
 
   void reset();
 
 private:
-  std::vector<Vec2<float>> initialMesh;
-  std::vector<Vec2<float>> oldMesh;
-  std::vector<Vec2<float>> mesh;
-  std::vector<Vec2<float>> newMesh;
+  std::vector<Vec2<double>> initialMesh;
+  std::vector<Vec2<double>> oldMesh;
+  std::vector<Vec2<double>> mesh;
+  std::vector<Vec2<double>> newMesh;
 };
